Cleanup + update submodule + fix MS / Big crashes
