+++ SADXCharSel/E100-Fixes.cpp
@@ -1,75 +1,63 @@
 #include "stdafx.h"
 
-Trampoline* E101_Main_t;
-Trampoline* MK2Main_t;
-Trampoline* E103_Main_t;
-Trampoline* E104_Main_t;
-
+static FunctionHook<void, task*> E101_Main_t((intptr_t)E101_Main);
+static FunctionHook<void, task*> E103_Main_t((intptr_t)E103_Main);
+static FunctionHook<void, task*> E104_Main_t((intptr_t)E104_Main);
+static FunctionHook<void, task*> MK2Main_t((intptr_t)MK2_Main);
 
 //Allow non gamma character to damage E100 Series.
-void CheckAndSetDamage(EntityData1* data1, EntityData1* p1)
+void CheckAndSetDamage(taskwk* data1, taskwk* p1)
 {
-	if (GetCollidingEntityA(data1) && p1->Status & Status_Attack && p1->CharID != Characters_Gamma && data1->Action > 0)
+	if (GetCollidingEntityA((EntityData1*)data1) && p1->flag & Status_Attack && p1->counter.b[1] != Characters_Gamma && data1->mode > 0)
 	{
-		data1->Status |= Status_Hurt;
+		data1->flag |= Status_Hurt;
 	}
 }
 
-void E104Enemy_Main_R(ObjectMaster* obj) {
+void E104Enemy_Main_R(task* obj) {
 
-	EntityData1* data1 = obj->Data1;
-	EntityData1* p1 = EntityData1Ptrs[0];
-
-	CheckAndSetDamage(data1, p1);
-
-	ObjectFunc(origin, E104_Main_t->Target());
-	origin(obj);
-}
-
-
-
-void E103Enemy_Main_R(ObjectMaster* obj) {
-
-	EntityData1* data1 = obj->Data1;
-	EntityData1* p1 = EntityData1Ptrs[0];
+	auto data1 = obj->twp;
+	auto p1 = playertwp[0];
 
 	CheckAndSetDamage(data1, p1);
 
-	ObjectFunc(origin, E103_Main_t->Target());
-	origin(obj);
+	E104_Main_t.Original(obj);
 }
 
+void E103Enemy_Main_R(task* obj) {
 
-void E101_Main_R(ObjectMaster* obj) {
-
-	EntityData1* data1 = obj->Data1;
-	EntityData1* p1 = EntityData1Ptrs[0];
+	auto data1 = obj->twp;
+	auto p1 = playertwp[0];
 
 	CheckAndSetDamage(data1, p1);
 
-	ObjectFunc(origin, E101_Main_t->Target());
-	origin(obj);
+	E103_Main_t.Original(obj);
 }
 
+void E101_Main_R(task* obj) {
 
-
-void MK2_Main_R(ObjectMaster* obj) {
-
-	EntityData1* data1 = obj->Data1;
-	EntityData1* p1 = EntityData1Ptrs[0];
+	auto data1 = obj->twp;
+	auto p1 = playertwp[0];
 
 	CheckAndSetDamage(data1, p1);
 
-	ObjectFunc(origin, MK2Main_t->Target());
-	origin(obj);
+	E101_Main_t.Original(obj);
 }
 
+void MK2_Main_R(task* obj) {
 
+	auto data1 = obj->twp;
+	auto p1 = playertwp[0];
+
+	CheckAndSetDamage(data1, p1);
+
+	MK2Main_t.Original(obj);
+}
 
 void __cdecl E100_Series_Fixes_Init()
 {
-	E103_Main_t = new Trampoline((int)E103_Main, (int)E103_Main + 0x7, E103Enemy_Main_R);
-	E104_Main_t = new Trampoline((int)E104_Main, (int)E104_Main + 0x7, E104Enemy_Main_R);
-	E101_Main_t = new Trampoline((int)E101_Main, (int)E101_Main + 0x5, E101_Main_R);
-	MK2Main_t = new Trampoline((int)MK2_Main, (int)MK2_Main + 0x5, MK2_Main_R);
+	E103_Main_t.Hook(E103Enemy_Main_R);
+	E104_Main_t.Hook(E104Enemy_Main_R);
+	E101_Main_t.Hook(E101_Main_R);
+	MK2Main_t.Hook(MK2_Main_R);
 }
