+++ SADXCharSel/mod.h
@@ -3,3 +3,7 @@
 extern HelperFunctions help;
 
 void Init_GammaFixes();
+void init_Patches();
+
+#define pNum obj->twp->counter.b[0]
+#define charID obj->twp->counter.b[1]
